Commiting AOL contributions to sipXtapi merge branch
git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/bandreasen-sipXtapi-AOL-merge@10722 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/include/rtcp/RTCPTimer.h
@@ -1,3 +1,20 @@
+// Copyright 2008 AOL LLC.
+// Licensed to SIPfoundry under a Contributor Agreement.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301
+// USA. 
 //
 // Copyright (C) 2004-2006 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
@@ -280,7 +297,7 @@ private:    // Private Methods
 
 #elif defined(__pingtel_on_posix__) /* ] [ */
 
-    static void ReportingAlarm(const int userData, const int eventData);
+    static void ReportingAlarm(const intptr_t userData, const intptr_t eventData);
 
 #endif /* ] */
