+++ sipXmediaLib/src/mp/MprRecorder.cpp
@@ -1,3 +1,20 @@
+// Copyright 2008 AOL LLC.
+// Licensed to SIPfoundry under a Contributor Agreement.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301
+// USA. 
 //  
 // Copyright (C) 2006-2007 SIPez LLC. 
 // Licensed to SIPfoundry under a Contributor Agreement. 
@@ -396,7 +413,7 @@ UtlBoolean MprRecorder::doProcessFrame(MpBufPtr inBufs[],
 
 void MprRecorder::progressReport(Completion code)
 {
-   int ud;
+   intptr_t ud;
 
    mStatus = code;
 
@@ -438,7 +455,7 @@ void MprRecorder::progressReport(Completion code)
                // the event was probably just reset, 
                // try again after waiting for 10 ms.
                OsTask::delay(10);
-               int userdata;
+               intptr_t userdata;
                mpEvent->getUserData(userdata);
                OsSysLog::add(FAC_MP, PRI_WARNING, 
                              "user data - old (0x%08x), "
