+++ sipXportLib/src/os/OsTimer.cpp
@@ -1,4 +1,21 @@
-// 
+// Copyright 2008 AOL LLC.
+// Licensed to SIPfoundry under a Contributor Agreement.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301
+// USA. 
+//  
 // Copyright (C) 2005-2007 SIPez LLC.
 // Licensed to SIPfoundry under a Contributor Agreement.
 // 
@@ -275,9 +292,9 @@ int OsTimer::getUserData()
    // Have to cast mpNotifier into OsQueuedEvent* to get the userData.
    OsQueuedEvent* e = dynamic_cast <OsQueuedEvent*> (mpNotifier);
    assert(e != 0);
-   int userData;
+   intptr_t userData;
    e->getUserData(userData);
-   return userData;
+   return (int) userData;
 }
 
 unsigned OsTimer::hash() const
